// frontend/src/sections/cold-chain-equipment/catalogue.generated.ts
export type AssetCategoriesListVariables = {
  filter?: {
    id?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    name?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    classId?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
  } | null;
};

export type AssetCategoriesListResult = {
  assetCategories: ({
  __typename: "AssetCategoryConnector";
} & {
  totalCount: number;
  nodes: Array<{
  id: string;
  name: string;
  classId: string;
}>;
});
};

export const AssetCategoriesList = {
  query: "query assetCategoriesList($filter: AssetCategoryFilterInput) {\n  assetCategories(filter: $filter) {\n    ... on AssetCategoryConnector {\n      __typename\n      totalCount\n      nodes {\n        id\n        name\n        classId\n      }\n    }\n  }\n}",
} as TypedDocument<AssetCategoriesListResult, AssetCategoriesListVariables>;

export type AssetTypesListVariables = {
  filter?: {
    id?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    name?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    categoryId?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
  } | null;
};

export type AssetTypesListResult = {
  assetTypes: ({
  __typename: "AssetTypeConnector";
} & {
  totalCount: number;
  nodes: Array<{
  id: string;
  name: string;
  categoryId: string;
}>;
});
};

export const AssetTypesList = {
  query: "query assetTypesList($filter: AssetTypeFilterInput) {\n  assetTypes(filter: $filter) {\n    ... on AssetTypeConnector {\n      __typename\n      totalCount\n      nodes {\n        id\n        name\n        categoryId\n      }\n    }\n  }\n}",
} as TypedDocument<AssetTypesListResult, AssetTypesListVariables>;

export type AssetCatalogueItemsListVariables = {
  filter?: {
    id?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    category?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    categoryId?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    class?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    classId?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    code?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    manufacturer?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    model?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    type?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    typeId?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    search?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    subCatalogue?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
  } | null;
  page?: {
    first?: number | null;
    offset?: number | null;
  } | null;
};

export type AssetCatalogueItemsListResult = {
  assetCatalogueItems: ({
  __typename: "AssetCatalogueItemConnector";
} & {
  totalCount: number;
  nodes: Array<{
  id: string;
  code: string;
  manufacturer: string | null;
  model: string;
  assetCategoryId: string;
  assetTypeId: string;
  assetType: {
  id: string;
  name: string;
} | null;
}>;
});
};

export const AssetCatalogueItemsList = {
  query: "query assetCatalogueItemsList($filter: AssetCatalogueItemFilterInput, $page: PaginationInput) {\n  assetCatalogueItems(filter: $filter, page: $page) {\n    ... on AssetCatalogueItemConnector {\n      __typename\n      totalCount\n      nodes {\n        id\n        code\n        manufacturer\n        model\n        assetCategoryId\n        assetTypeId\n        assetType {\n          id\n          name\n        }\n      }\n    }\n  }\n}",
} as TypedDocument<AssetCatalogueItemsListResult, AssetCatalogueItemsListVariables>;